Improve coverage on tests with babel
TypeScript compiled to ES6 doesn't change much, so we now test ES6 code
so that coverage report is a bit more accurate. Added 'babel-jest',
'babel-preset-es2015' and 'babel-preset-react' as dev dependencies as
well as a '.babelrc' file so that jest can transpile the ES6 code and
report coverage.
Added Enzyme as dev dependency to use the shallow renderer so that we can
actually test that Link is being used.

Updated readme with install instructions